% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{cell_counts}
\alias{cell_counts}
\alias{cell_counts.estimatecc}
\alias{cell_counts,estimatecc-method}
\title{cell_counts}
\usage{
cell_counts(object)

\S4method{cell_counts}{estimatecc}(object)

\S4method{cell_counts}{estimatecc}(object)
}
\arguments{
\item{object}{a \code{estimatecc} object}
}
\description{
Accessors for the 'cell_counts' slot of a estimatecc object.
}
