% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{cellcounts}
\alias{cellcounts}
\alias{cellcounts}
\alias{cellcounts.estimateCC}
\alias{cellcounts,estimateCC-method}
\alias{cellcounts,estimateCC-method}
\title{cellcounts}
\usage{
cellcounts(object)

\S4method{cellcounts}{estimateCC}(object)

\S4method{cellcounts}{estimateCC}(object)
}
\arguments{
\item{object}{a \code{estimateCC} object}

\item{...}{other}
}
\description{
Accessors for the 'cellcounts' slot of a estimateCC object.
}
