% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{celltypeSpecificDMRegions}
\alias{celltypeSpecificDMRegions}
\title{Celltype specific differentially methylated regions}
\format{add more here.}
\description{
This is the script used to create the
\code{celltypeSpecificDMRegions} data set.
The purpose is use in the \code{estimateCC()}
function.

The object was created using the script in /inst and
located in the /data folder.
}
