% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.extract_raw_data}
\alias{.extract_raw_data}
\title{Extract raw data}
\usage{
.extract_raw_data(object)
}
\arguments{
\item{object}{an object can be a \code{RGChannelSet}, 
\code{GenomicMethylSet} or \code{BSseq} object}
}
\value{
A list preprocessed objects from the 
\code{RGChannelSet}, \code{GenomicMethylSet} or 
\code{BSseq} objects to be used in \code{.preprocess_estimatecc()}.
}
\description{
Extract the methylation values and 
GRanges objects
}
