% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.initialize_theta}
\alias{.initialize_theta}
\title{.initialize_theta}
\usage{
.initialize_theta(n, K, alpha0 = NULL, alpha1 = NULL, sig0 = NULL,
  sig1 = NULL, tau = NULL)
}
\arguments{
\item{n}{Number of samples}

\item{K}{Number of cell types}

\item{alpha0}{Default NULL. Initial mean methylation level in unmethylated regions}

\item{alpha1}{Default NULL. Initial mean methylation level in methylated regions}

\item{sig0}{Default NULL. Initial var methylation level in unmethylated regions}

\item{sig1}{Default NULL. Initial var methylation level in methylated regions}

\item{tau}{Default NULL. Initial var for measurement error}
}
\description{
Creates a container with initial theta
parameter estimates
}
