% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.initializeMLEs}
\alias{.initializeMLEs}
\title{.initializeMLEs}
\usage{
.initializeMLEs(init_param_method, n, K, Ys, Zs, a0init, a1init, sig0init,
  sig1init, tauinit)
}
\arguments{
\item{init_param_method}{Add here.}

\item{n}{Add here.}

\item{K}{Add here.}

\item{Ys}{Add here.}

\item{Zs}{Add here.}

\item{a0init}{Add here.}

\item{a1init}{Add here.}

\item{sig0init}{Add here.}

\item{sig1init}{Add here.}

\item{tauinit}{Add here.}
}
\description{
Helper functions to initialize 
MLEs in estimatecc.
}
