% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.methylcc_engine}
\alias{.methylcc_engine}
\title{.methylcc_engine}
\usage{
.methylcc_engine(Ys, Zs, current_pi_mle, current_theta, epsilon, max_iter)
}
\arguments{
\item{Ys}{observed methylation levels in samples provided by user
of dimension R x n}

\item{Zs}{Cell type specific regions of dimension R x K}

\item{current_pi_mle}{cell composition MLE estimates of dimension K x n}

\item{current_theta}{other parameter estimates in EM algorithm}

\item{epsilon}{Add here.}

\item{max_iter}{Add here.}
}
\description{
Helper function for estimatecc
}
