% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.methylcc_estep}
\alias{.methylcc_estep}
\title{Expectation step}
\usage{
.methylcc_estep(Ys, Zs, current_pi_mle, current_theta, meth_status = 0)
}
\arguments{
\item{Ys}{observed methylation levels in samples provided by user
of dimension R x n}

\item{Zs}{Cell type specific regions of dimension R x K}

\item{current_pi_mle}{cell composition MLE estimates of dimension K x n}

\item{current_theta}{other parameter estimates in EM algorithm}

\item{meth_status}{Indicator function corresponding to regions that
are unmethylated (meth_status=0) or methylated (meth_status=1)}
}
\value{
List of expected value of the first two moments 
of the random effects (or the E-Step in the EM algorithm)
used in \code{.methylcc_engine()}
}
\description{
Expectation step in EM algorithm for methylCC
}
