% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.methylcc_mstep}
\alias{.methylcc_mstep}
\title{Maximization step}
\usage{
.methylcc_mstep(Ys, Zs, current_pi_mle, current_theta, estep0, estep1)
}
\arguments{
\item{Ys}{observed methylation levels in samples provided by user
of dimension R x n}

\item{Zs}{Cell type specific regions of dimension R x K}

\item{current_pi_mle}{cell composition MLE estimates of dimension K x n}

\item{current_theta}{other parameter estimates in EM algorithm}

\item{estep0}{Results from expectation step for unmethylated regions}

\item{estep1}{Results from expectation step for methylated regions}
}
\value{
A list of the updated MLEs (or the M-Step 
in the EM algorithm) used in \code{.methylcc_engine()}
}
\description{
Maximization step in EM Algorithm for methylCC
}
