% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.pick_target_positions}
\alias{.pick_target_positions}
\title{.pick_target_positions}
\usage{
.pick_target_positions(target_granges, target_object = NULL,
  target_cvg = NULL, dmp_regions)
}
\arguments{
\item{target_granges}{add more here.}

\item{target_object}{an optional argument which contains the meta-data for targetRanges. If
targetRanges already contains the meta-data, do not need to supply \code{target_object}.}

\item{target_cvg}{add more here.}

\item{dmp_regions}{add more here.}
}
\description{
Pick probes from target data using the indices in \code{dmp_regions}.
}
