% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.preprocess_estimatecc}
\alias{.preprocess_estimatecc}
\title{.preprocess_estimatecc}
\usage{
.preprocess_estimatecc(object, verbose = TRUE,
  init_param_method = "random",
  celltype_specific_dmrs = celltype_specific_dmrs)
}
\arguments{
\item{object}{an object can be a \code{RGChannelSet}, 
\code{GenomicMethylSet} or \code{BSseq} object}

\item{verbose}{TRUE/FALSE argument specifying if verbose
messages should be returned or not. Default is TRUE.}

\item{init_param_method}{method to initialize parameter estimates.
Choose between "random" (randomly sample) or "known_regions"
(uses unmethyalted and methylated regions that were identified
based on Reinus et al. (2012) cell sorted data.).
Defaults to "random".}

\item{celltype_specific_dmrs}{cell type specific differentially 
methylated regions (DMRs).}
}
\value{
A list of object to be used in estimatecc
}
\description{
This function preprocesses the data 
before the \code{estimatecc()} function
}
