% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.splitit}
\alias{.splitit}
\title{.splitit}
\usage{
.splitit(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A list to be used in \code{find_dmrs()}
}
\description{
helper function to split along a variable
}
