% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.WFun}
\alias{.WFun}
\title{Helper function to take the product of Z and
cell composition estimates}
\usage{
.WFun(Zs, pi_mle)
}
\arguments{
\item{Zs}{Cell type specific regions of dimension R x K}

\item{pi_mle}{cell composition MLE estimates}
}
\value{
A list of output after taking the product of Z 
and cell composition mle estimates to be used in 
\code{.methylcc_estep()}.
}
\description{
Helper function which is the product of
Z and pi_mle
}
