% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{estimatecc-class}
\alias{estimatecc-class}
\title{the estimatecc class}
\value{
\code{summary} returns the summary information 
about the cell composition estimate procedure and 
\code{cell_counts} returns the cell composition estimates
}
\description{
Objects of this class store all 
the values needed information to work with a
estimatecc object
}
\section{Slots}{

\describe{
\item{\code{summary}}{information about the samples and 
regions used to estimate cell composition}

\item{\code{cell_counts}}{cell composition estimates}
}}

\examples{
\dontrun{
library(FlowSorted.Blood.450k)
data(FlowSorted.Blood.450k)
rgset <- FlowSorted.Blood.450k[,
     pData(FlowSorted.Blood.450k)$CellTypeLong \%in\% "Whole blood"]
est <- estimatecc(object = rgset) 
}

}
