% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{estimateCC}
\alias{estimateCC}
\title{estimateCC}
\usage{
estimateCC(object, regionMat = NULL, verbose = TRUE, epsilon = 0.01,
  maxIter = 50, initParamMethod = "random", includeSE = FALSE, B = 0,
  a0init = NULL, a1init = NULL, sig0init = NULL, sig1init = NULL)
}
\arguments{
\item{object}{an object can be an \code{RGChannelSet} or
\code{BSseq} object. The \code{object} can also be a
data frame or matrix with observations (e.g. CpGs)
on the rows and samples as the columns.}

\item{regionMat}{If the \code{object} is not an
\code{RGChannelSet} or \code{BSseq} object, then
the user must provide \code{regionMat}. 
This must have the same number of
rows as \code{object} (number of regions or CpGs)
and one column for each cell type. The values in
\code{regionMat} should be either 0 or 1 representing
regions (or CpGs) that are either methylated (1) or
not methylated (0).}

\item{verbose}{TRUE/FALSE argument specifying if verbose
messages should be returned or not. Default is TRUE.}

\item{epsilon}{Threshold for Expectation-Maximization
(EM) algorithm}

\item{maxIter}{Maximum number of iterations for EM
algorithm}

\item{initParamMethod}{add more here}

\item{includeSE}{TRUE/FALSE argument specifying if
standard error estimates should be included. Default
is FALSE The parameter \code{B} must be set to a value
greater than 0 if \code{includeSE == TRUE}}

\item{B}{Number of bootstrap samples to simulate to
calculate SE estimates.}

\item{a0init}{add more here.}

\item{a1init}{add more here.}

\item{sig0init}{add more here.}

\item{sig1init}{add more here.}
}
\description{
Estimate cell composition from seq data
}
