% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{estimateCCPlot}
\alias{estimateCCPlot}
\title{Plot results from \code{estimateCC} function.}
\usage{
estimateCCPlot(objectEst, savePlot = FALSE)
}
\arguments{
\item{objectEst}{an estimateCC object from \code{estimateCC}}

\item{savePlot}{a TRUE/FALSE object argument
determining if the plot will be saved for further use or
immediately displayed on the screen.}
}
\value{
The cell composition estimates for each sample.
If \code{includeSE} argument was set to TRUE in
\code{estimateCC()}, then the standard errors will
also be plotted.
}
\description{
This function plots the parameter estimates
and standard error estimates from the \code{estimateCC}
function.
}
