% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dmrs.R
\docType{methods}
\name{find_dmrs}
\alias{find_dmrs}
\title{find_dmrs}
\usage{
find_dmrs(verbose = TRUE, gr_target = NULL, include_cpgs = FALSE,
  include_dmrs = TRUE, num_cpgs = 50, num_regions = 50,
  bumphunter_beta_cutoff = 0.2, dmr_up_cutoff = 0.6,
  dmr_down_cutoff = 0.5, dmr_pval_cutoff = 1e-11,
  cpg_pval_cutoff = 1e-08, cpg_up_dm_cutoff = 0,
  cpg_down_dm_cutoff = 0, pairwise_comparison = FALSE,
  mset_train_flow_sort = NULL)
}
\arguments{
\item{verbose}{Add here.}

\item{gr_target}{Add here.}

\item{include_cpgs}{Add here.}

\item{include_dmrs}{Add here.}

\item{num_cpgs}{Add here.}

\item{num_regions}{Add here.}

\item{bumphunter_beta_cutoff}{Add here.}

\item{dmr_up_cutoff}{Add here.}

\item{dmr_down_cutoff}{Add here.}

\item{dmr_pval_cutoff}{Add here.}

\item{cpg_pval_cutoff}{Add here.}

\item{cpg_up_dm_cutoff}{Add here.}

\item{cpg_down_dm_cutoff}{Add here.}

\item{pairwise_comparison}{Add here.}

\item{mset_train_flow_sort}{Add here.}
}
\value{
Add here.
}
\description{
Add here.
}
\details{
Add here.
}
\examples{
found_regions <- find_dmrs()

}
