% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{methylCC_EStep}
\alias{methylCC_EStep}
\alias{methylCC_Estep}
\title{Expectation step}
\usage{
methylCC_Estep(ZMat, pi.mle, theta, YMat, methStatus = 0)
}
\arguments{
\item{ZMat}{Cell type specific regions of dimension R x K}

\item{pi.mle}{cell composition MLE estimates of dimension K x n}

\item{theta}{other parameter estimates in EM algorithm}

\item{YMat}{observed methylation levels in samples provided by user
of dimension R x n}

\item{methStatus}{Indicator function corresponding to regions that
are unmethylated (methStatus=0) or methylated (methStatus=1)}
}
\description{
Expectation step in EM algorithm for methylCC
}
