% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{methylCC_MStep}
\alias{methylCC_MStep}
\title{Maximization step}
\usage{
methylCC_MStep(ZMat, pi.mle, theta, YMat, EStep0, EStep1)
}
\arguments{
\item{ZMat}{Cell type specific regions of dimension R x K}

\item{pi.mle}{cell composition MLE estimates of dimension K x n}

\item{theta}{other parameter estimates in EM algorithm}

\item{YMat}{observed methylation levels in samples provided by user
of dimension R x n}

\item{EStep0}{Results from expectation step for unmethylated regions}

\item{EStep1}{Results from expectation step for methylated regions}
}
\description{
Maximization step in EM Algorithm for methylCC
}
