% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{mleList}
\alias{mleList}
\alias{mleList}
\alias{mleList.estimateCC}
\alias{mleList,estimateCellCountSeq-method}
\alias{mleList,estimateCC-method}
\title{mleList}
\usage{
mleList(object)

\S4method{mleList}{estimateCC}(object)

\S4method{mleList}{estimateCC}(object)
}
\arguments{
\item{object}{a \code{estimateCC} object}

\item{...}{other}
}
\description{
Accessors for the 'mleList' slot of a estimateCC object.
}
