% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{mles}
\alias{mles}
\alias{mles}
\alias{mles.estimateCC}
\alias{mles,estimateCellCountSeq-method}
\alias{mles,estimateCC-method}
\title{mles}
\usage{
mles(object)

\S4method{mles}{estimateCC}(object)

\S4method{mles}{estimateCC}(object)
}
\arguments{
\item{object}{a \code{estimateCC} object}

\item{...}{other}
}
\description{
Accessors for the 'mles' slot of a estimateCC object.
}
