% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{MStep}
\alias{MStep}
\title{Maximization step}
\usage{
MStep(ZMat, est, YMat, R, n, ids, del0, del1)
}
\arguments{
\item{ZMat}{add more here}

\item{est}{add more here}

\item{YMat}{add more here}

\item{R}{add more here}

\item{n}{add more here}

\item{ids}{add more here}

\item{del0}{add more here}

\item{del1}{add more here}
}
\description{
Maximization step in EM Algorithm
}
