% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{pickTargetPositions}
\alias{pickTargetPositions}
\title{pickTargetPositions}
\usage{
pickTargetPositions(targetgRanges, targetObject = NULL, targetCVG = NULL,
  dmpRegions)
}
\arguments{
\item{targetgRanges}{add more here.}

\item{targetObject}{an optional argument which contains the meta-data for targetRanges. If
targetRanges already contains the meta-data, do not need to supply targetObject.}

\item{targetCVG}{add more here.}

\item{dmpRegions}{add more here.}
}
\description{
Pick probes from target data using the indices in dmpRegions.
}
