% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{preProcessEstCountsSeq}
\alias{preProcessEstCountsSeq}
\title{preProcessEstCountsSeq}
\usage{
preProcessEstCountsSeq(object, regionMat, verbose = TRUE, initParamMethod)
}
\arguments{
\item{object}{an object can be an \code{RGChannelSet} or
\code{BSseq} object. The \code{object} can also be a
data frame or matrix with observations (e.g. CpGs)
on the rows and samples as the columns.}

\item{regionMat}{If the \code{object} is not an
\code{RGChannelSet} or \code{BSseq} object, then
the user must provide \code{regionMat}. 
This must have the same number of
rows as \code{object} (number of regions or CpGs)
and one column for each cell type. The values in
\code{regionMat} should be either 0 or 1 representing
regions (or CpGs) that are either methylated (1) or
not methylated (0).}

\item{verbose}{TRUE/FALSE argument specifying if verbose
messages should be returned or not. Default is TRUE.}

\item{initParamMethod}{add more here}
}
\description{
add more here
}
