% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary.estimatecc}
\alias{summary,estimatecc-method}
\title{summary}
\usage{
summary(object, ...)

\S4method{summary}{estimatecc}(object)

\S4method{summary}{estimatecc}(object)
}
\arguments{
\item{object}{a \code{estimatecc} object}
}
\description{
Accessors for the 'summary' slot of a estimatecc object.
}
