% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCCFunctions.R
\name{WFun}
\alias{WFun}
\title{Helper function to take the product of Z and
cell composition estimates}
\usage{
WFun(ZMat, pi.mle)
}
\arguments{
\item{ZMat}{Cell type specific regions of dimension R x K}

\item{pi.mle}{cell composition MLE estimates}
}
\description{
Helper function which is the product of
Z and pi.mle
}
