% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{cell_counts}
\alias{cell_counts}
\alias{cell_counts,estimatecc-method}
\title{Generic function that returns the cell composition 
estimates}
\usage{
cell_counts(object)

\S4method{cell_counts}{estimatecc}(object)
}
\arguments{
\item{object}{an object of class \code{estimatecc}.}
}
\value{
Returns the cell composition estimates
}
\description{
Given a estimatecc object, this function returns the 
cell composition estimates

Accessors for the 'cell_counts' slot of a estimatecc object.
}
\examples{
\dontrun{
library(FlowSorted.Blood.450k)
data(FlowSorted.Blood.450k)
rgset <- FlowSorted.Blood.450k[,
     pData(FlowSorted.Blood.450k)$CellTypeLong \%in\% "Whole blood"]
est <- estimatecc(object = rgset) 
cell_counts(est)
}

}
