% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.initializeMLEs}
\alias{.initializeMLEs}
\title{.initializeMLEs}
\usage{
.initializeMLEs(init_param_method, n, K, Ys, Zs, a0init, a1init, sig0init,
  sig1init, tauinit)
}
\arguments{
\item{init_param_method}{method to initialize parameter estimates.
Choose between "random" (randomly sample) or "known_regions"
(uses unmethyalted and methylated regions that were identified
based on Reinus et al. (2012) cell sorted data.).
Defaults to "random".}

\item{n}{Number of samples}

\item{K}{Number of cell types}

\item{Ys}{observed methylation levels in samples 
provided by user of dimension R x n}

\item{Zs}{Cell type specific regions of dimension R x K}

\item{a0init}{Default NULL. Initial mean methylation 
level in unmethylated regions}

\item{a1init}{Default NULL. Initial mean methylation
level in methylated regions}

\item{sig0init}{Default NULL. Initial var methylation 
level in unmethylated regions}

\item{sig1init}{Default NULL. Initial var methylation 
level in methylated regions}

\item{tauinit}{Default NULL. Initial var for measurement error}
}
\value{
A list of MLE estimates to be used in \code{estimatecc()}.
}
\description{
Helper functions to initialize 
MLEs in \code{estimatecc()}.
}
