% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.pick_target_positions}
\alias{.pick_target_positions}
\title{Pick target positions}
\usage{
.pick_target_positions(target_granges, target_object = NULL,
  target_cvg = NULL, dmp_regions)
}
\arguments{
\item{target_granges}{add more here.}

\item{target_object}{an optional argument which 
contains the meta-data for \code{target_granges}. If
\code{target_granges} already contains the meta-data, do not 
need to supply \code{target_object}.}

\item{target_cvg}{coverage reads for the target object}

\item{dmp_regions}{differentially methylated regions}
}
\value{
A list of GRanges objects to be used in 
\code{.preprocess_estimatecc()}
}
\description{
Pick probes from target data using 
the indices in \code{dmp_regions}
}
