% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatecc.R
\name{estimatecc}
\alias{estimatecc}
\title{estimatecc}
\usage{
estimatecc(object, find_dmrs_object = NULL, region_mat = NULL,
  verbose = TRUE, epsilon = 0.01, max_iter = 100,
  init_param_method = "random", a0init = NULL, a1init = NULL,
  sig0init = NULL, sig1init = NULL, tauinit = NULL,
  mset_train_flow_sort = NULL, use_houseman_cpgs = FALSE)
}
\arguments{
\item{object}{an object can be an \code{RGChannelSet} or
\code{BSseq} object. The \code{object} can also be a
data frame or matrix with observations (e.g. CpGs)
on the rows and samples as the columns.}

\item{find_dmrs_object}{Add here.}

\item{region_mat}{If the \code{object} is not an
\code{RGChannelSet} or \code{BSseq} object, then
the user must provide \code{region_mat}.
This must have the same number of
rows as \code{object} (number of regions or CpGs)
and one column for each cell type. The values in
\code{region_mat} should be either 0 or 1 representing
regions (or CpGs) that are either methylated (1) or
not methylated (0).}

\item{verbose}{TRUE/FALSE argument specifying if verbose
messages should be returned or not. Default is TRUE.}

\item{epsilon}{Threshold for EM algorithm to check
for convergence. Default is 0.01.}

\item{max_iter}{Maximum number of iterations for EM
algorithm. Default is 100 iterations.}

\item{init_param_method}{method to initialize parameter estimates.
Choose between "random" (randomly sample) or "known_regions"
(uses unmethyalted and methylated regions that were identified
based on Reinus et al. (2012) cell sorted data.).
Defaults to "random".}

\item{a0init}{Default NULL. Initial mean 
methylation level in unmethylated regions}

\item{a1init}{Default NULL. Initial mean 
methylation level in methylated regions}

\item{sig0init}{Default NULL. Initial var 
methylation level in unmethylated regions}

\item{sig1init}{Default NULL. Initial var 
methylation level in methylated regions}

\item{tauinit}{Default NULL. Initial var 
for measurement error}

\item{mset_train_flow_sort}{Add here.}

\item{use_houseman_cpgs}{Add here.}
}
\description{
Estimate cell composition from seq data
}
