% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatecc.R
\docType{methods}
\name{estimatecc}
\alias{estimatecc}
\title{Estimate cell composition from DNAm data}
\usage{
estimatecc(object, find_dmrs_object = NULL, verbose = TRUE,
  epsilon = 0.01, max_iter = 100, take_intersection = FALSE,
  init_param_method = "random", a0init = NULL, a1init = NULL,
  sig0init = NULL, sig1init = NULL, tauinit = NULL)
}
\arguments{
\item{object}{an object can be a \code{RGChannelSet}, 
\code{GenomicMethylSet} or \code{BSseq} object}

\item{find_dmrs_object}{If the user would like to supply 
different differentially methylated regions, they can 
use the output from the \code{find_dmrs} function 
to supply different regions to \code{estimatecc}.}

\item{verbose}{TRUE/FALSE argument specifying if verbose
messages should be returned or not. Default is TRUE.}

\item{epsilon}{Threshold for EM algorithm to check
for convergence. Default is 0.01.}

\item{max_iter}{Maximum number of iterations for EM
algorithm. Default is 100 iterations.}

\item{take_intersection}{TRUE/FALSE asking if only the CpGs 
included in \code{object} should be used to find DMRs. 
Default is FALSE.}

\item{init_param_method}{method to initialize parameter estimates.
Choose between "random" (randomly sample) or "known_regions"
(uses unmethyalted and methylated regions that were identified
based on Reinus et al. (2012) cell sorted data.).
Defaults to "random".}

\item{a0init}{Default NULL. Initial mean 
methylation level in unmethylated regions}

\item{a1init}{Default NULL. Initial mean 
methylation level in methylated regions}

\item{sig0init}{Default NULL. Initial var 
methylation level in unmethylated regions}

\item{sig1init}{Default NULL. Initial var 
methylation level in methylated regions}

\item{tauinit}{Default NULL. Initial var 
for measurement error}
}
\value{
A object of the class \code{estimatecc} that 
contains information about the cell composition 
estimation (in the \code{summary} slot) and 
the cell composition estimates themselves 
(in the \code{cell_counts} slot).
}
\description{
Estimate cell composition from DNAm data
}
\examples{
library(FlowSorted.Blood.450k)
data(FlowSorted.Blood.450k)
rgset <- FlowSorted.Blood.450k[,
     pData(FlowSorted.Blood.450k)$CellTypeLong \%in\% "Whole blood"]
est <- estimatecc(object = rgset) 

}
