% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary}
\alias{summary.estimateCC}
\alias{summary,estimateCC-method}
\alias{summary,estimateCC-method}
\title{summary}
\usage{
summary(object, ...)

\S4method{summary}{estimateCC}(object)

\S4method{summary}{estimateCC}(object)
}
\arguments{
\item{object}{a \code{estimateCC} object}
}
\description{
Accessors for the 'summary' slot of a estimateCC object.
}
